test/fixtures/wpt/performance-timeline/tentative/include-frames-originA-A.html

<!DOCTYPE html>

<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<body>
<script>
  const loadSameOriginChildFrame = () => {
    return new Promise(resolve => {

      const childFrame = document.createElement('iframe');

      childFrame.addEventListener("load", () => {
        resolve();
      });

      childFrame.src = "../resources/child-frame.html";

      document.body.appendChild(childFrame);
    });
  }

  promise_test(async () => {
    performance.clearResourceTimings();

    // Load a child frame.
    await loadSameOriginChildFrame();

    const entries = performance.getEntries({ includeChildFrames: true });

    const navigationEntries = performance.getEntries({ entryType: "navigation", includeChildFrames: true });

    const markedEntries = performance.getEntries(
      { name: 'entry-name', includeChildFrames: true });

    // 3 entries for parent, 2 for child.
    assert_equals(entries.length, 5, 'Total entries should be 5.');

    // 1 entry for parent, 1 for child.
    assert_equals(navigationEntries.length, 2, 'Navigation entries should be 2.');

    // 1 entry for child.
    assert_equals(markedEntries.length, 1, 'Mark entries should be 1.');

    // Test cases where includeChildFrames is false.
    const entriesWithNoFitlerOptions = performance.getEntries();

    const entriesWithoutIncludingChildFrames = performance.getEntries({ includeChildFrames: false });

    const navigationEntriesWithoutIncludingChildFrames = performance.getEntries({ entryType: "navigation", includeChildFrames: false });

    const markedEntriesWithoutIncludingChildFrames = performance.getEntries(
      { name: 'entry-name', includeChildFrames: false });

    // 3 entries for parent.
    assert_equals(entriesWithNoFitlerOptions.length, 3,
      'Total entries without filter options should be 3 without filter options.');

    assert_equals(entriesWithoutIncludingChildFrames.length, 3,
      'Total entries with includeChildFrame being false should be 3.');

    // 1 entry for parent.
    assert_equals(navigationEntriesWithoutIncludingChildFrames.length, 1,
      'Navigation entries with includeChildFrame being false should be 1.');

    // 0 entry for child.
    assert_equals(markedEntriesWithoutIncludingChildFrames.length, 0,
      'Mark entries with includeChildFrame being false should be 0.');

  }, 'GetEntries of a document of origin A and its child frame of origin A.');
</script>
</body>
